Put the CV at the front of the list of plugins so Cutter always picks it up when I check cuts only on the first universe in each group.  Cutter was keeping some sums of weights as ints.  They're doubles now which means AnaTool efficiency isn't 0.

+++ cuts/Cutter.h
@@ -45,10 +45,10 @@ namespace util
       double totalWeightPassed() const; //Get total weight that passed all cuts
 
     private:
-      long int fSumAllAnaToolWeights;
-      long int fSumSignalTruthWeights;
-      long int fSumSignalAnaToolWeights;
-      long int fSumAllTruthWeights;
+      double fSumAllAnaToolWeights;
+      double fSumSignalTruthWeights;
+      double fSumSignalAnaToolWeights;
+      double fSumAllTruthWeights;
       long int fNRecoEntries;
 
       reco_t fRecoPreCuts;
